profiles: Remove network mode option
It's currently broken in both Lineage 15.1 and 16.0,
and requires changes all over the place, which are prone
to introduce more damage than the benefit that the feature
really holds and no one seems to be able to finish.
Remove the option to avoid dealing with the constant
bug reports of it crashing.

Change-Id: Ia3897c46732766d7a790880516a34fdb26f7a568